[refactoring] img 폴더 생성
