+++ client/src/component/Schedule/Schedule.scss
@@ -318,11 +318,11 @@
       position: fixed;
       right: 3% !important;
       .kakao {
-        background-image: url("../../kakaobank_app.png");
+        background-image: url("../img/kakaobank_app.png");
         background-color: rgb(254, 228, 9) !important;
       }
       .toss {
-        background-image: url("../../toss_app.png");
+        background-image: url("../img/toss_app.png");
         background-color: rgb(246, 246, 246) !important;
       }
       .won {
