+++ client/src/component/Signin/Signin.scss
@@ -109,7 +109,7 @@
       width : 100%;
       button{
         box-shadow: none !important;
-        background-image: url("../../googleLogo.png");
+        background-image: url("../img/googleLogo.png");
         background-position: 30% 50%;
         background-size: 36px;
         background-repeat: no-repeat;
@@ -137,7 +137,7 @@
             display: block;
             width: 400px;
             margin: 10px auto;
-            background-image: url("../../googleLogo.png");
+            background-image: url("../img/googleLogo.png");
             background-position: 25% 50%;
             background-size: 36px;
             background-repeat: no-repeat;
@@ -157,7 +157,7 @@
           display: block;
           width: 400px;
           margin: 10px auto;
-          background-image: url("../../facebookLogo.png");
+          background-image: url("../img/facebookLogo.png");
           background-position: 25% 50%;
           background-size: 36px;
           background-repeat: no-repeat;
@@ -367,7 +367,7 @@
           width : 100%;
           button{
             box-shadow: none;
-            background-image: url("../../googleLogo.png");
+            background-image: url("../img/googleLogo.png");
             background-position: 24% 50%;
             background-size: 30px;
             background-repeat: no-repeat;
@@ -394,7 +394,7 @@
                 display: block;
                 width: 300px;
                 margin: 10px auto;
-                background-image: url("../../googleLogo.png");
+                background-image: url("../img/googleLogo.png");
                 background-position: 18% 50%;
                 background-size: 30px;
                 background-repeat: no-repeat;
@@ -414,7 +414,7 @@
               display: block;
               width: 300px;
               margin: 10px auto;
-              background-image: url("../../facebookLogo.png");
+              background-image: url("../img/facebookLogo.png");
               background-position: 18% 50%;
               background-size: 30px;
               background-repeat: no-repeat;
