+++ client/src/component/img/googleLogo.png
